supabase/requirements.txt
supabase==2.24.0
python-dotenv==1.0.0
